pyproject.toml
[project]
name = "seafront"
version = "2.6.1"
requires-python = ">=3.13"
authors = [{ name = "Patrick Hennig", email = "patrick.hennig@uu.se" }]
dependencies = [
    "gxipy @ git+https://github.com/slaide/daheng-imaging-gxipy@v1.0.0", # galaxy camera API
    "toupcam @ git+https://github.com/slaide/toupcam-sdk@v1.0.3", # touptek camera API
    "seaconfig @ git+https://github.com/slaide/seafront-config@v6.3.0", # configuration file handling
    "pyserial==3.5", # for microcontroller communication
    "crc>=1.3,<1.4", # for basic microcontroller packet communication
    "pillow>=11.1,<11.2", # for image format conversion to store to disk/send to client for display
    "scipy>=1.15,<1.16", # for laser aufocus system signal processing
    "numpy>=2,<3", # for signal processing and various math
    "tifffile==2025.1.10", # for tiff image format
    "imagecodecs==2024.12.30", # for image compression
    "matplotlib>=3,<4", # for debug functionality
    "fastapi==0.115.8", # http server
    "uvicorn[standard]>=0.34,<0.35", # http server with threading support
    "websockets>=14.2,<14.3", # to enable uvicorn websocket capabilities
    "pydantic>=2.9,<2.10", # python type model verification
    "ome-types", # for ome info in image files
    "opencv-python>=4.11,<4.12", # fast gaussian blur for noise autofocus images
    "loguru==0.7.3", # logging
    "json5", # better json file formatting
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0", # type checker
    "ruff>=0.3.0",    # linter
    "torch>=2.10.0",
    "types-pillow",   # type stubs for pillow
    "types-pyserial", # type stubs for pyserial
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["seafront"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportMissingTypeStubs = true
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportImplicitStringConcatenation = true
reportUnreachableCode = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",  # line length violations
    "C901",  # function complexity too high
    "N801",  # class name should use CapWords
    "N802",  # function name should be lowercase
    "N803",  # argument name should be lowercase
    "N806",  # variable name should be lowercase
    "N815",  # variable name should not be mixed case
    "E722",  # do not use bare except
    "UP024", # replace aliased errors
    "W293",  # blank line contains whitespace
    "W291",  # trailing whitespace
    "I001",  # import block is unsorted
    "F541",  # f-string without any placeholders
    "RUF022",# <x> is not sorted (??)
]

[tool.ruff.lint.isort]
known-first-party = ["seafront"]

[dependency-groups]
dev = [
    "pyright>=1.1.402",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.setuptools.package-data]
toupcam = ["web-static"]
